Update Js to check if tab open
It seems that the blur function is really the only reason to check cheating and it's not working well since you can't really violate someone's privacy and see if they are opening tabs and this function just cries if they try and click away. Maybe I can have it time out instead? After a minute of practicing their word, while the judge makes comments in the background?